Improve makefile

+++ Makefile
@@ -3,20 +3,27 @@ BIN = node_modules/.bin
 all: install
 
 install: build
-	cp ./example-config.json config.json
-	@echo "[i] Install complete."
+	if ! [ -e ./config.json ]
+	then
+		@echo "[i] No previous config file, using example-config..."
+		cp ./example-config.json config.json
+	else
+		@echo "[!] Old configuration file detected, keeping it"
+		cp ./example-config.json config.json.new
+		@echo "[i] New version of config file saved as 'config.json.new'"
+	fi
+	@echo "[i] Install complete"
 
 test: build
 	@echo "[i] Running tests..."
 	@npm run test
-	@echo "[i] Tests complete."
+	@echo "[i] Tests complete"
 
 clean:
 	@echo "[i] Clearing generated files..."
 	rm -rf ./logs
 	rm -rf ./public/js/cognactime.min.js
 	rm -rf ./public/css/cognactime.min.css
-	rm -rf ./config.json
 	@make install
 
 prod:
@@ -24,7 +31,7 @@ prod:
 	@echo "Fetching remote git tree and resetting..."
 	@git fetch && git reset --hard origin/master
 	@make clean
-	@echo "Not set up for prod yet."
+	@echo "Cognac Time!"
 
 dev: build
 	@echo "[i] Starting development server..."
@@ -33,11 +40,11 @@ dev: build
 build: node_modules
 	@echo "[i] Triggering build..."
 	@npm run build
-	@echo "[i] Build complete."
+	@echo "[i] Build complete"
 
 node_modules:
 	@echo "[i] Fetching dependencies..."
 	@npm install
-	@echo "[i] Completed fetching dependencies."
+	@echo "[i] Completed fetching dependencies"
 
 .PHONY: test clean prod dev node_modules
